:necktie: Add useCustomInfiniteQuery reusable hook
